+++ recordset.go
@@ -41,6 +41,8 @@ type Recordset struct {
 
 	active    *AtomicBool
 	cancelled chan struct{}
+
+	chanLock sync.Mutex
 }
 
 // NewRecordset generates a new RecordSet instance.
@@ -62,24 +64,24 @@ func (rcs *Recordset) IsActive() bool {
 	return rcs.active.Get()
 }
 
-// Results returns a new receive-only channel with the results of the Scan/Query
+// Results returns a new receive-only channel with the results of the Scan/Query.
 // This is a more idiomatic approach to the iterator pattern in getting the
 // results back from the recordset, and doesn't require the user to write the
 // ugly select in their code.
-// Result embeds A Record and an error reference.
+// Result contains a Record and an error reference.
 //
 // Example:
 //
-// recordset, err := client.ScanAll(nil, namespace, set)
-// handleError(err)
-// for res := range recordset.Results() {
-//   if res.Err != nil {
-//     // handle error here
-//   } else {
-//     // process record here
-//     fmt.Println(res.Record.Bins)
-//   }
-// }
+//  recordset, err := client.ScanAll(nil, namespace, set)
+//  handleError(err)
+//  for res := range recordset.Results() {
+//    if res.Err != nil {
+//      // handle error here
+//    } else {
+//      // process record here
+//      fmt.Println(res.Record.Bins)
+//    }
+//  }
 func (rcs *Recordset) Results() <-chan *Result {
 	res := make(chan *Result, len(rcs.Records))
 
@@ -115,6 +117,8 @@ func (rcs *Recordset) Close() {
 		// wait till all goroutines are done
 		rcs.wgGoroutines.Wait()
 
+		rcs.chanLock.Lock()
+		defer rcs.chanLock.Unlock()
 		close(rcs.Records)
 		close(rcs.Errors)
 	}
@@ -124,5 +128,13 @@ func (rcs *Recordset) signalEnd() {
 	rcs.wgGoroutines.Done()
 	if rcs.goroutines.DecrementAndGet() == 0 {
 		rcs.Close()
+	}
+}
+
+func (rcs *Recordset) sendError(err error) {
+	rcs.chanLock.Lock()
+	defer rcs.chanLock.Unlock()
+	if rcs.IsActive() {
+		rcs.Errors <- err
 	}
 }